Moves tests to same package as the tested class.
