+++ embabel-agent-test-support/embabel-agent-test/src/test/java/com/embabel/agent/test/integration/ExampleAgentApplication.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.agent.test.integration.example;
+package com.embabel.agent.test.integration;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
