+++ embabel-agent-test-support/embabel-agent-test/src/test/java/com/embabel/agent/test/integration/WriteAndReviewAgentIntegrationTest.java
@@ -13,46 +13,40 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.agent.test.integration.example;
+package com.embabel.agent.test.integration;
 
 import com.embabel.agent.api.invocation.AgentInvocation;
 import com.embabel.agent.domain.io.UserInput;
-import com.embabel.agent.test.integration.EmbabelMockitoIntegrationTest;
-import com.embabel.agent.test.integration.example.StreamingWriteAndReviewAgent.StreamingReviewedStory;
-import org.junit.jupiter.api.BeforeEach;
+import com.embabel.agent.test.integration.WriteAndReviewAgent.ReviewedStory;
 import org.junit.jupiter.api.Test;
-import reactor.core.publisher.Flux;
 
-import static com.embabel.agent.test.integration.example.StreamingWriteAndReviewAgent.StreamingStory;
-import static org.junit.jupiter.api.Assertions.*;
+import static com.embabel.agent.test.integration.WriteAndReviewAgent.Story;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Use framework superclass to test the complete workflow of writing and reviewing a story.
  * This will run under Spring Boot against an AgentPlatform instance
  * that has loaded all our agents.
  */
-class StreamingWriteAndReviewAgentIntegrationTest extends EmbabelMockitoIntegrationTest {
-
-    @BeforeEach
-    void enableStreaming(){
-        supportsStreaming(true);
-    }
+class WriteAndReviewAgentIntegrationTest extends EmbabelMockitoIntegrationTest {
 
     @Test
     void shouldExecuteCompleteWorkflow() {
         var input = new UserInput("Write about artificial intelligence");
 
-        var story = new StreamingStory("AI will transform our world...");
-        var reviewedStory = new StreamingReviewedStory(story, "Excellent exploration of AI themes.", Personas.REVIEWER);
+        var story = new Story("AI will transform our world...");
+        var reviewedStory = new ReviewedStory(story, "Excellent exploration of AI themes.", Personas.REVIEWER);
 
-        whenCreateObjectStream(s -> s.contains("Craft a short story"), StreamingStory.class)
-                .thenReturn(Flux.just(story));
+        whenCreateObject(s -> s.contains("Craft a short story"), Story.class)
+                .thenReturn(story);
 
         // The second call uses generateText
-        whenGenerateStream(s -> s.contains("You will be given a short story to review"))
-                .thenReturn(Flux.just(reviewedStory.review()));
+        whenGenerateText(s -> s.contains("You will be given a short story to review"))
+                .thenReturn(reviewedStory.review());
 
-        var invocation = AgentInvocation.create(agentPlatform, StreamingReviewedStory.class);
+        var invocation = AgentInvocation.create(agentPlatform, ReviewedStory.class);
         var reviewedStoryResult = invocation.invoke(input);
 
         assertNotNull(reviewedStoryResult);
@@ -61,9 +55,9 @@ class StreamingWriteAndReviewAgentIntegrationTest extends EmbabelMockitoIntegrat
         assertEquals(reviewedStory, reviewedStoryResult,
                 "Expected review to match: " + reviewedStoryResult);
 
-        verifyCreateObjectStreamMatching(prompt -> prompt.contains("Craft a short story"), StreamingStory.class,
+        verifyCreateObjectMatching(prompt -> prompt.contains("Craft a short story"), Story.class,
                 llm -> llm.getLlm().getTemperature() != null && llm.getLlm().getTemperature() == 0.9 && llm.getToolGroups().isEmpty());
-        verifyGenerateStreamMatching(prompt -> prompt.contains("You will be given a short story to review"));
+        verifyGenerateTextMatching(prompt -> prompt.contains("You will be given a short story to review"));
         verifyNoMoreInteractions();
     }
 }
